requirements.txt
aiofiles==0.5.0
aniso8601==7.0.0
asgiref==3.4.1
asttokens==2.0.4
async-exit-stack==1.0.1
async-generator==1.10
bcrypt==3.2.0
beautifulsoup4==4.9.3
blis==0.7.4
catalogue==2.0.6
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
cryptography==3.4.6
cycler==0.11.0
cymem==2.0.5
decorator==4.4.2
dnspython==2.1.0
docker==4.4.3
ecdsa==0.14.1
el-core-news-lg @ https://github.com/explosion/spacy-models/releases/download/el_core_news_lg-3.2.0/el_core_news_lg-3.2.0-py3-none-any.whl
email-validator==1.1.2
en-core-web-lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.2.0/en_core_web_lg-3.2.0-py3-none-any.whl
english==2020.7.0
executing==0.5.4
fastapi==0.63.0
fasttext==0.9.2
fonttools==4.33.3
graphene==2.1.8
graphql-core==2.3.2
graphql-relay==2.0.1
graphviz==0.16
gunicorn==20.1.0
h11==0.12.0
httptools==0.1.1
icecream==2.1.0
idna==2.10
importlib-metadata==3.4.0
itsdangerous==1.1.0
Jinja2==2.11.3
joblib==1.1.0
kiwisolver==1.3.2
langcodes==3.3.0
MarkupSafe==1.1.1
matplotlib==3.5.1
monotonic==1.5
motor==2.3.1
murmurhash==1.0.5
neo4j==4.4.1
neotime==1.7.4
networkx==2.6.3
numpy==1.20.1
orjson==3.4.8
packaging==20.9
pandas==1.3.5
pansi==2020.7.3
passlib==1.7.4
pathy==0.4.0
Pillow==9.0.0
preshed==3.0.5
promise==2.3
prompt-toolkit==2.0.10
py2neo==2021.0.0
pyasn1==0.4.8
pybind11==2.9.2
pycparser==2.20
pydantic==1.8.2
Pygments==2.8.0
pymongo==3.11.3
pyparsing==2.4.7
pytextrank==3.2.2
python-dateutil==2.8.2
python-decouple==3.4
python-dotenv==0.15.0
python-jose==3.2.0
python-multipart==0.0.5
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
rsa==4.7.1
Rx==1.6.1
schedule==1.2.0
scikit-learn==1.0.2
scikit-learn-extra==0.2.0
scipy==1.7.3
six==1.15.0
smart-open==3.0.0
soupsieve==2.2
spacy==3.2.1
spacy-legacy==3.0.8
spacy-loggers==1.0.1
srsly==2.4.2
starlette==0.13.6
thinc==8.0.13
threadpoolctl==3.1.0
tqdm==4.57.0
typer==0.3.2
typing-extensions==3.7.4.3
ujson==3.2.0
urllib3==1.26.3
uvicorn==0.13.4
uvloop==0.15.2
wasabi==0.8.2
watchgod==0.7
wcwidth==0.2.5
websocket-client==0.57.0
websockets==8.1
yellowbrick==1.4
zipp==3.4.0
